refactor: move morpho directory to within tools
